+++ returns/context/requires_context_result.py
@@ -6,7 +6,7 @@ from typing_extensions import final
 
 from returns.context.requires_context import RequiresContext
 from returns.primitives.container import BaseContainer
-from returns.primitives.types import Immutable, Stateless
+from returns.primitives.types import Immutable
 from returns.result import Failure, Result, Success
 
 # Context:
@@ -649,11 +649,7 @@ class RequiresContextResult(
 
 
 @final
-class ContextResult(
-    Immutable,
-    Stateless,
-    Generic[_EnvType],
-):
+class ContextResult(Immutable, Generic[_EnvType]):
     """
     Helpers that can be used to work with ``RequiresContextResult`` container.